This change is necessary to make remix-auth work with the shopify oauth flow for admin apps.
The reason is, the remix app will be installed in different stores with different URLs. This means the authorization and access token url changes from user to user.
Since the strategy is a single instance for the whole app, this was the best method I found.

First I extended the oauth2 strategy and basically overrode the authorizationURL and the tokenURL for each request. This however could lead to race conditions, where one request sets the auth url to one shop and at the same time another one sets it to a different shop url.

So I kinda abused the context to pass the authURl + tokenURL dynamically. For shopify this is not a problem because every time the app is started the shop name is in the searchParams of the URL.

Also important, the tokenURL and authorizationURL can still be set in the init, but if the urls are passed in the context it is preferred.

Please let me know what you think of it :)