Allow nullable parameters for `ConstructorOptions` (#142)
This PR modifies the `ConstructorOptions` class to support nullable
values for `clientSecret` and `redirectURI`.

This adjustment is crucial for aligning with the operational logic of
the Artic OAuth2 library, where certain configurations may not require
these fields. The change primarily affects the
`validateAuthorizationCode` method, enhancing its flexibility to handle
different authentication scenarios (e.g. `if (this.clientPassword ===
null) { ...`).